+++ src/store/keplr/actions.js
@@ -1,4 +1,8 @@
 import { CONFIG } from '@/constants';
+import {
+  assertIsDeliverTxSuccess,
+  SigningStargateClient,
+} from '@cosmjs/stargate';
 
 export default {
   async connect({ commit, dispatch }, { translator, context }) {
@@ -19,6 +23,7 @@ export default {
     commit('setLoading', true);
     await dispatch('suggestChain', { chain, $t });
     await dispatch('getAccounts', { chain, $t });
+    dispatch('subscribeKeyStoreChange', { chain, $t });
     commit('setLoading', false);
   },
   async suggestChain({ commit }, { chain, $t }) {
@@ -57,6 +62,11 @@ export default {
             coinMinimalDenom: CONFIG.STABLE_COIN.DENOM,
             coinDecimals: CONFIG.STABLE_COIN.EXPONENT,
           },
+          {
+            coinDenom: CONFIG.TOKEN.SYMBOL,
+            coinMinimalDenom: CONFIG.TOKEN.DENOM,
+            coinDecimals: CONFIG.TOKEN.EXPONENT,
+          },
         ],
         stakeCurrency: {
           coinDenom: CONFIG.TOKEN.SYMBOL,
@@ -85,6 +95,37 @@ export default {
       commit('setError', $t('msgs.noAccountFound'));
     }
   },
+  subscribeKeyStoreChange({ commit, dispatch }, { chain, $t }) {
+    window.addEventListener('keplr_keystorechange', async () => {
+      commit('setLoading', true);
+      commit('setAccounts', []);
+      commit('setInitialized', false);
+      await dispatch('getAccounts', { chain, $t });
+      commit('setLoading', false);
+    });
+  },
+  async signAndBroadcastTransaction({ commit, getters }, { msgs }) {
+    try {
+      const chain = CONFIG.CHAIN.LIST.find(
+        (item) => item.lcd === process.env.VUE_APP_LCD
+      );
+      const offlineSigner = window.keplr.getOfflineSigner(chain.chainId);
+      const client = await SigningStargateClient.connectWithSigner(
+        chain.rpc,
+        offlineSigner
+      );
+      const accounts = getters['accounts'];
+      const fee = _calcFee(msgs.length);
+      const result = await client.signAndBroadcast(
+        accounts[0].address,
+        msgs,
+        fee
+      );
+      assertIsDeliverTxSuccess(result);
+    } catch (error) {
+      commit('setError', error);
+    }
+  },
   disconnect({ commit }) {
     commit('reset');
   },
@@ -92,3 +133,13 @@ export default {
     commit('setError', undefined);
   },
 };
+
+const _calcFee = (msgs) => ({
+  amount: [
+    {
+      denom: CONFIG.STABLE_COIN.DENOM,
+      amount: (CONFIG.FEE_AMOUNT * msgs).toString(),
+    },
+  ],
+  gas: (CONFIG.GAS_AMOUNT * msgs).toString(),
+});