+++ CHANGELOG.md
@@ -9,7 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Add proposal voting by Keplr
+- Add transaction execution functions in the chain using Keplr:  
+  1. Delegating tokens
+  2. Undelegating tokens
+  3. Redelegating tokens
+  4. Claim tokens
+  5. Voting on proposals
 - Add proposal stats
 
 ### Changed